refactor: 清理遗留中间件、废弃 UI 与历史表，为 8 大 Plan 实施建立基线
删除 5 个无调用方中间件（DeferredToolFilter/LoopDetection/SubagentLimit/ThreadData/Title）

删除 ACPServer 及 4 个废弃 UI 页面（StartupInit/NotificationCenter/OnDeviceModel+VM/TaskList）

MIGRATION_10_11 中 memories/memories_fts/task_queue 由 CREATE 改为 DROP（表已废弃）

同步清理引用方代码，净减 2107 行

+++ app/src/main/kotlin/com/lin/hippyagent/core/task/HippyJobJson.kt
@@ -30,7 +30,7 @@ object HippyJobJson {
 
     private fun anyToJsonElement(value: Any): JsonElement = when (value) {
         is String -> JsonPrimitive(value)
-        is Int -> JsonPrimitive(value)
+        is Int -> JsonPrimitive(value.toLong())
         is Long -> JsonPrimitive(value)
         is Double -> JsonPrimitive(value)
         is Float -> JsonPrimitive(value)
